+++ packages/components-angular/tests/vrt/specs/tag.vrt.ts
@@ -1,5 +1,5 @@
 import { defaultViewports, getVisualRegressionTester, vrtTest } from '@porsche-design-system/shared/testing';
 
 it.each(defaultViewports)('should have no visual regression for viewport %s', async (viewport) => {
-  expect(await vrtTest(getVisualRegressionTester(viewport), 'tag-status', '/tag-status')).toBeFalsy();
+  expect(await vrtTest(getVisualRegressionTester(viewport), 'tag', '/tag')).toBeFalsy();
 });