+++ packages/components/scripts/cleanReadmes.ts
@@ -26,7 +26,7 @@ const adjustHeadline = (str: string): string => {
     'p-inline-notification': 'Notifications',
     'p-table': 'Table',
     'p-tabs': 'Tabs',
-    'p-tag-status': 'Tags',
+    'p-tag': 'Tags',
   };
 
   // all component names on multi prop pages