+++ packages/storefront/storefront.config.ts
@@ -510,19 +510,19 @@ export const config: StorefrontConfig = {
       ],
     },
     Tags: {
-      'Tag Status': [
+      Tag: [
         // @ts-ignore
-        () => import('@/../../components/src/components/action/tags/tag-status/tag-status.examples.md'),
+        () => import('@/../components/src/components/action/tag/tag.examples.md'),
       ],
       'Tag Dismissible': [
         // @ts-ignore
-        () => import('@/../../components/src/components/action/tags/tag-dismissible/tag-dismissible.examples.md'),
+        () => import('@/../../components/src/components/action/tag-dismissible/tag-dismissible.examples.md'),
       ],
       Props: [
         // @ts-ignore
-        () => import('@/../../components/src/components/action/tags/tag-status/tag-status.props.md'),
+        () => import('@/../components/src/components/action/tag/tag.props.md'),
         // @ts-ignore
-        () => import('@/../../components/src/components/action/tags/tag-dismissible/tag-dismissible.props.md'),
+        () => import('@/../../components/src/components/action/tag-dismissible/tag-dismissible.props.md'),
       ],
     },
     'Text Field': {