+++ packages/components/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- `Headline` `hyphens` property which can be set to `none`
+
 ### [2.9.3-rc.0-skeletons] - 2022-03-29
 
 ### [2.9.3-rc.0] - 2022-03-28