+++ packages/components/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - `Tag Status`
+- `Tag Dismissible`
 
 ### [2.9.3-rc.1] - 2022-04-06