+++ packages/components/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 #### Added
-- `Tag Status`
+- `Tag`
 - `Tag Dismissible`
 
 ### [2.9.3-rc.1] - 2022-04-06